Corrected code that raise error after numpy update
When using the library, numpy raised an error relating to the deprecation of np.int. It suggests that the new correct way of doing this is using int() or any of the numpy integer functions (such as np.int32, np.int64, etc.). I propose using int() because, upon cursory testing of the function, I found it to solve the issue wile also being retrocompatible.